build: :wrench: styles outsourced in seperate file
